Rename 'error' package to 'duerror' in utils directory
Updated the package name in multiple files to resolve potential naming conflicts with the standard 'error' package. This ensures better clarity and maintains consistency across the codebase.
