+++ app/backend/utils/duerror/DUError.go
@@ -1,4 +1,4 @@
-package error
+package duerror
 
 type DUError interface {
 	Error() string
