+++ app/backend/utils/duerror/connection.go
@@ -1,4 +1,4 @@
-package error
+package duerror
 
 type ConnectionError struct {
 	msg string
