+++ app/backend/utils/duerror/fileIO.go
@@ -1,4 +1,4 @@
-package error
+package duerror
 
 type FileIOError struct {
 	msg string
