+++ app/backend/utils/duerror/memoryFull.go
@@ -1,4 +1,4 @@
-package error
+package duerror
 
 type MemoryFullError struct {
 	msg string
