+++ app/backend/utils/duerror/send.go
@@ -1,4 +1,4 @@
-package error
+package duerror
 
 type SendError struct {
 	msg string
